refactor: 순수 자바 테스트를 위해 Interface로 JpaRepository 감싸기 - 기존 레포는 fake 구현시 사용하지 않는 많은 매소드를 오버라이딩 해줘야 함
