+++ dongmanee/src/main/java/com/dongmanee/domain/club/dao/ClubCategoryDataRepository.java
@@ -1,0 +1,9 @@
+package com.dongmanee.domain.club.dao;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.dongmanee.domain.club.domain.ClubCategory;
+import com.dongmanee.domain.club.service.port.ClubCategoryRepository;
+
+public interface ClubCategoryDataRepository extends JpaRepository<ClubCategory, Long>, ClubCategoryRepository {
+}
