--- dongmanee/src/main/java/com/dongmanee/domain/club/dao/ClubCategoryRepository.java
@@ -1,9 +1,0 @@
-package com.dongmanee.domain.club.dao;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.dongmanee.domain.club.domain.ClubCategory;
-
-public interface ClubCategoryRepository
-	extends JpaRepository<ClubCategory, Long> {
-}
