+++ dongmanee/src/main/java/com/dongmanee/domain/club/dao/ClubDataRepository.java
@@ -3,6 +3,7 @@ package com.dongmanee.domain.club.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.dongmanee.domain.club.domain.Club;
+import com.dongmanee.domain.club.service.port.ClubRepository;
 
-public interface ClubRepository extends JpaRepository<Club, Long> {
+public interface ClubDataRepository extends JpaRepository<Club, Long>, ClubRepository {
 }
