+++ dongmanee/src/main/java/com/dongmanee/domain/club/dao/ClubSnsDataRepository.java
@@ -1,0 +1,9 @@
+package com.dongmanee.domain.club.dao;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.dongmanee.domain.club.domain.ClubSns;
+import com.dongmanee.domain.club.service.port.ClubSnsRepository;
+
+public interface ClubSnsDataRepository extends JpaRepository<ClubSns, Long>, ClubSnsRepository {
+}
