--- dongmanee/src/main/java/com/dongmanee/domain/club/dao/ClubSnsRepository.java
@@ -1,8 +1,0 @@
-package com.dongmanee.domain.club.dao;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.dongmanee.domain.club.domain.ClubSns;
-
-public interface ClubSnsRepository extends JpaRepository<ClubSns, Long> {
-}
