+++ dongmanee/src/main/java/com/dongmanee/domain/club/service/port/ClubCategoryRepository.java
@@ -1,0 +1,9 @@
+package com.dongmanee.domain.club.service.port;
+
+import java.util.Optional;
+
+import com.dongmanee.domain.club.domain.ClubCategory;
+
+public interface ClubCategoryRepository {
+	Optional<ClubCategory> findById(Long id);
+}
