+++ dongmanee/src/main/java/com/dongmanee/domain/club/service/port/ClubRepository.java
@@ -1,0 +1,7 @@
+package com.dongmanee.domain.club.service.port;
+
+import com.dongmanee.domain.club.domain.Club;
+
+public interface ClubRepository {
+	Club save(Club club);
+}
