+++ dongmanee/src/main/java/com/dongmanee/domain/club/service/port/ClubSnsRepository.java
@@ -1,0 +1,13 @@
+package com.dongmanee.domain.club.service.port;
+
+import java.util.Optional;
+
+import com.dongmanee.domain.club.domain.ClubSns;
+
+public interface ClubSnsRepository {
+	ClubSns save(ClubSns clubSns);
+
+	Optional<ClubSns> findById(Long id);
+
+	void delete(ClubSns clubSns);
+}
