+++ dongmanee/src/main/java/com/dongmanee/domain/club/service/port/ClubUserRepository.java
@@ -1,0 +1,11 @@
+package com.dongmanee.domain.club.service.port;
+
+import java.util.Optional;
+
+import com.dongmanee.domain.club.domain.ClubUser;
+
+public interface ClubUserRepository {
+	ClubUser save(ClubUser clubUser);
+
+	Optional<ClubUser> findClubUserWithMemberClub(Long memberId, Long clubId);
+}
